Using a multiplier to keep combo values integer

+++ src/utils/combo.h
@@ -14,8 +14,8 @@
 
 typedef int           boolean; /* logical variable         */
 typedef int           ntype;   /* number of states/items   */
-typedef float         itype;   /* item profits and weights */
-typedef float         stype;   /* sum of profit or weight  */
+typedef long          itype;   /* item profits and weights */
+typedef long          stype;   /* sum of profit or weight  */
 
 /* item record */
 typedef struct {
